% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_tool.R
\name{coor_pull}
\alias{coor_pull}
\title{Query min value of Start and max value of End}
\usage{
coor_pull(
  tfmapperdb = tfmapperdb,
  species_key = species_key,
  ip_key = ip_key,
  chr_key = chr_key
)
}
\arguments{
\item{tfmapperdb}{default `tfmapperdb`, SQLiteConnection name}

\item{species_key}{default `species_key`, keywords of Species}

\item{ip_key}{default `ip_key`, keywords of IP}

\item{chr_key}{default `chr_key`, keywords of Chr}
}
\value{
possible coordination
}
\description{
Query min value of Start and max value of End
}
\examples{
\dontrun{
coor_pull()
}
}
