% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_tool.R
\name{merge_query}
\alias{merge_query}
\title{Merge gene data from `query_gene` or `query_coord` and annotation data from `query_anno`}
\usage{
merge_query(gene_data, anno_info)
}
\arguments{
\item{gene_data}{data from `query_gene` or `query_coord`}

\item{anno_info}{data from `query_anno`}
}
\value{
a data frame contains results of query
}
\description{
Merge gene data from `query_gene` or `query_coord` and annotation data from `query_anno`
}
\examples{
\dontrun{
merge_query(gene_pos,gene_anno)
}
}
