% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_tool.R
\name{query_anno}
\alias{query_anno}
\title{Query annotation information of gene data from `query_gene` or `query_coord`}
\usage{
query_anno(
  tfmapperdb = tfmapperdb,
  gene_data,
  cell_line_key = cell_line_key,
  cell_type_key = cell_type_key,
  tissue_type_key = tissue_type_key
)
}
\arguments{
\item{tfmapperdb}{default `tfmapperdb`, SQLiteConnection name}

\item{gene_data}{result from query_gene or query_coord}

\item{cell_line_key}{default `cell_line_key`, keywords of Cell line}

\item{cell_type_key}{default `cell_type_key`, keywords of Cell type}

\item{tissue_type_key}{default `tissue_type_key`, keywords of tissue type}
}
\value{
a data.frame contains annotation results
}
\description{
Query annotation information of gene data from `query_gene` or `query_coord`
}
\examples{
\dontrun{
query_anno(gene_pos)
}
}
