% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_tool.R
\name{query_coord}
\alias{query_coord}
\title{Query gene data from coordination of Start and End in a Chr}
\usage{
query_coord(
  tfmapperdb = tfmapperdb,
  species_key = species_key,
  chr_key = chr_key,
  ip_key = ip_key,
  start_key = start_key,
  end_key = end_key
)
}
\arguments{
\item{tfmapperdb}{default `tfmapperdb`, SQLiteConnection name}

\item{species_key}{default `species_key`, keywords of Species}

\item{chr_key}{default `chr_key`, keywords of Chr}

\item{ip_key}{default `ip_key`, keywords of IP}

\item{start_key}{default `start_key`, coordination of Start}

\item{end_key}{default `end_key`, coordination of End}
}
\value{

}
\description{
Query gene data from coordination of Start and End in a Chr
}
\examples{
\dontrun{
query_coord()
}
}
