% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_tool.R
\name{query_gene}
\alias{query_gene}
\title{Query gene data according to gene name}
\usage{
query_gene(
  tfmapperdb = tfmapperdb,
  species_key = species_key,
  ip_key = ip_key,
  portion_key = portion_key,
  gene_key = gene_key
)
}
\arguments{
\item{tfmapperdb}{default `tfmapperdb`, SQLiteConnection name}

\item{species_key}{default `species_key`, keywords of Species}

\item{ip_key}{default `ip_key`, keywords of IP}

\item{portion_key}{default `portion_key`, keywords of Portion}

\item{gene_key}{default `gene_key`, keywords of  Gene}
}
\value{
data.frame tbl contains gene data
}
\description{
Query gene data according to gene name
}
\examples{
\dontrun{
query_gene()
}
}
